WIP: sharding
Enable partitioning of a document namespace into disjoint sub
namespaces, aka "shards".

A client specifies the value, aka "key-value", of the sharding key
using an operation's query param or a body field. The key-value is
mapped to a shard id. A shard is mapped to a set of nodes from the
nodegroup. The number of nodes is the replicationFactor, as the shard
is a replication unit. An owner is selected from the shard's nodes.

Mapping of key-values to the same shard as well as sharing a host
between two different shards can be allowed or not, based on config.

A ShardsManagementService has been added to create and manage shards.
There is a single instance of ShardsManagementService per NodeSelector
that supports sharding. That instance is called: "shards manager".

The default node selector, implemented by ConsistentHashingNodeSelector,
supports sharding (replicationFactor != null). It works with its
shard manager. It caches the shards' map for quick, local ownership
and replication targets determination.

+++ xenon-common/src/test/java/com/vmware/xenon/services/common/TestNodeGroupService.java
@@ -267,7 +267,7 @@ public class TestNodeGroupService {
     private List<URI> expectedFailedHosts = new ArrayList<>();
     private String replicationTargetFactoryLink = ExampleService.FACTORY_LINK;
     private String replicationNodeSelector = ServiceUriPaths.DEFAULT_NODE_SELECTOR;
-    private long replicationFactor;
+    private int replicationFactor;
     private int replicationQuorum = 1;
 
     private Map<String, URI> replicationTargetLinks;
@@ -2798,7 +2798,7 @@ public class TestNodeGroupService {
 
     @Test
     public void replication1x() throws Throwable {
-        this.replicationFactor = 1L;
+        this.replicationFactor = 1;
         this.replicationNodeSelector = ServiceUriPaths.DEFAULT_1X_NODE_SELECTOR;
         this.replicationTargetFactoryLink = Replication1xExampleFactoryService.SELF_LINK;
         doReplication();
@@ -2806,7 +2806,7 @@ public class TestNodeGroupService {
 
     @Test
     public void replication3x() throws Throwable {
-        this.replicationFactor = 3L;
+        this.replicationFactor = 3;
         this.replicationNodeSelector = ServiceUriPaths.DEFAULT_3X_NODE_SELECTOR;
         this.replicationTargetFactoryLink = Replication3xExampleFactoryService.SELF_LINK;
         this.nodeCount = Math.max(5, this.nodeCount);
